dex: cleanup
